+++ backend/src/test/java/com/back/domain/product/product/controller/ApiV1ProductControllerTest.java
@@ -16,8 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
@@ -32,97 +31,8 @@ public class ApiV1ProductControllerTest {
     private ProductService productService;
 
     @Test
-    @DisplayName("상품 등록")
-    public void t1() throws Exception {
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/products")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                        {
-                                            "name": "상품 test",
-                                            "price": 10000,
-                                            "description": "상품 test",
-                                            "stock": 10
-                                        }
-                                        """)
-                ).andDo(print());
-
-        Product product = productService.findLatest().get();
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1ProductController.class))
-                .andExpect(handler().methodName("create"))
-                .andExpect(status().isCreated())
-                .andExpect(jsonPath("$.resultCode").value("201-1"))
-                .andExpect(jsonPath("$.msg").value("%d번 상품이 등록되었습니다.".formatted(product.getId())))
-                .andExpect(jsonPath("$.data.id").value(product.getId()))
-                .andExpect(jsonPath("$.data.createDate").value(Matchers.startsWith(product.getCreateDate().toString().substring(0, 20))))
-                .andExpect(jsonPath("$.data.modifyDate").value(Matchers.startsWith(product.getModifyDate().toString().substring(0, 20))))
-                .andExpect(jsonPath("$.data.name").value(product.getName()))
-                .andExpect(jsonPath("$.data.price").value(product.getPrice()))
-                .andExpect(jsonPath("$.data.description").value(product.getDescription()))
-                .andExpect(jsonPath("$.data.stock").value(product.getStock()));
-    }
-
-    @Test
-    @DisplayName("상품 등록 - without name")
-    public void t1_1() throws Exception {
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/products")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                        {
-                                            "name": "",
-                                            "price": 10000,
-                                            "description": "상품 test",
-                                            "stock": 10
-                                        }
-                                        """)
-                ).andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1ProductController.class))
-                .andExpect(handler().methodName("create"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.resultCode").value("400-1"))
-                .andExpect(jsonPath("$.msg").value("""
-                        name-NotBlank-must not be blank
-                        name-Size-size must be between 2 and 100
-                        """.stripIndent().trim()));
-    }
-
-    @Test
-    @DisplayName("상품 등록 - without price")
-    public void t1_2() throws Exception {
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/products")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                        {
-                                            "name": "상품 test",
-                                            "price": 0,
-                                            "description": "상품 test",
-                                            "stock": 10
-                                        }
-                                        """)
-                ).andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1ProductController.class))
-                .andExpect(handler().methodName("create"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.resultCode").value("400-1"))
-                .andExpect(jsonPath("$.msg").value("""
-                        price-Min-must be greater than or equal to 100
-                        """.stripIndent().trim()));
-    }
-
-    @Test
     @DisplayName("상품 다건 조회")
-    public void t2() throws Exception {
+    public void t1() throws Exception {
         ResultActions resultActions = mvc
                 .perform(
                         get("/api/v1/products")
@@ -152,7 +62,7 @@ public class ApiV1ProductControllerTest {
 
     @Test
     @DisplayName("상품 단건 조회")
-    public void t3() throws Exception {
+    public void t2() throws Exception {
         int id = 1;
 
         ResultActions resultActions = mvc
@@ -178,7 +88,7 @@ public class ApiV1ProductControllerTest {
 
     @Test
     @DisplayName("상품 단건 조회 - 404")
-    public void t3_1() throws Exception {
+    public void t2_1() throws Exception {
         int id = Integer.MAX_VALUE;
 
         ResultActions resultActions = mvc
@@ -193,59 +103,5 @@ public class ApiV1ProductControllerTest {
                 .andExpect(status().isNotFound())
                 .andExpect(jsonPath("$.resultCode").value("404-1"))
                 .andExpect(jsonPath("$.msg").value("해당 데이터가 존재하지 않습니다."));
-    }
-
-
-    @Test
-    @DisplayName("상품 삭제")
-    public void t4() throws Exception {
-        int id = 1;
-
-        ResultActions resultActions = mvc
-                .perform(
-                        delete("/api/v1/products/%d".formatted(id))
-                                .contentType(MediaType.APPLICATION_JSON)
-                ).andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1ProductController.class))
-                .andExpect(handler().methodName("delete"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.resultCode").value("200-1"))
-                .andExpect(jsonPath("$.msg").value("%d번 상품이 삭제되었습니다.".formatted(id)));
-    }
-
-    @Test
-    @DisplayName("상품 수정")
-    public void t5() throws Exception {
-        int id = 1;
-
-        ResultActions resultActions = mvc
-                .perform(
-                        put("/api/v1/products/%d".formatted(id))
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                        {
-                                            "name": "상품 new",
-                                            "price": 1000,
-                                            "description": "상품 new",
-                                            "stock": 100
-                                        }
-                                        """)
-                ).andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1ProductController.class))
-                .andExpect(handler().methodName("update"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.resultCode").value("200-1"))
-                .andExpect(jsonPath("$.msg").value("%d번 상품이 수정되었습니다.".formatted(id)));
-
-        Product product = productService.findById(id).get();
-
-        assertThat(product.getName()).isEqualTo("상품 new");
-        assertThat(product.getPrice()).isEqualTo(1000);
-        assertThat(product.getDescription()).isEqualTo("상품 new");
-        assertThat(product.getStock()).isEqualTo(100);
     }
 }
